product-search-service/requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
scikit-learn
scipy